ON-14969: Split ImageStream from Kustomization
Move ImageStream manifests into the new namespace, "onload-artifacts",
managed outside Kustomize.

Kustomize provides resource name transformations with the namePrefix and
nameSuffix fields. However, this does not play well with the ImageStream
consumers. If Kustomize changes the ImageStream name, it won't update
ImageStreamTag, e.g. in BuildConfig, or the FROM statements in
Dockerfiles. Consequently, these dependent resources won't be able to
use the renamed ImageStream resources.

To address this, there are two general approaches: implementing the
ImageStream name changes in the consumers or moving ImageStream resource
management away from Kustomize. This patch chooses the latter, assuming
that (1) Kustomize is more of a deployment-time tool and (2) we
currently do not distinguish between build-time and deployment-time in
this repository. If we chose the former approach, we would try applying
Kustomize to the build-time workflow, potentially outside of Kustomize's
scope.