ON-14968: Create SA for Onload Module
Create a privileged service account specifically for the Onload module,
so it doesn't rely on those provided with the "openshift-kmm" namespace.
Additionally, because Module is a custom CRD, add a new name reference
for Kustomize.

This patch makes the Onload Module movable in an arbitrary namespace and
it is a prerequisite for the upcoming Kustomization patch.