ON-14920: Fix SELinux context for Onload files
Set the "container_file_t" type for all files in "/opt/onload" in the
compute nodes from the device plugin DS.

The device plugin DS copies Onload files from the UL build image into
the node filesystem, ready for use in the user pods when they request
the relevant resources with "resources.limits.amd.com/onload: 1".
However, SELinux refuses to grant access to these files to unprivileged
pods because they have a wrong "type" in the SELinux
"user:role:type:level". The pods have the "container_file_t" type, but
the Onload files have the "var_t" type.

To address the mismatch, the device plugin DS runs "chcon" to set the
type in "/opt/onload" recursively.

Note that the "chcon" changes are not permanent and will be reset in the
next autorelabel with SELinux. This patch operates on the assumption
that autorelabel is a heavyweight operation and occurs only as part of
node maintenance, i.e., outside our DS's lifespan. If we discover that
this assumption is wrong, we should consider using "semanage fcontext"
for permanent changes. It may require a more involved fix, like a new
MachineConfig because it implies changes to the node system files,
i.e. in "/etc/selinux/".