cleanup: Trigger device plugin build automatically
Add build triggers to the Onload device plugin.

We couldn't anonymously pull the device plugin sources from the internet
until we published this repository. Instead, we (temporarily) asked the
user to trigger the device plugin build locally/manually with "oc
start-build ... --from-dir ...". Now, we can pull the sources, so this
patch instantiates the triggers.

Because the device plugin image is based on the Onload UL build image,
the trigger specifically refers to the onload-user ImageStreamTag to
retrigger the build as soon as the new Onload UL image is available.

When all Onload resources are created simultaneously, e.g. with
Kustomize, the UL build may not be available, and it may cause the early
device plugin builds to fail, but once UL is ready, the build restarts
and succeeds.