ON-15008: Allow the use of onload profiles
Onload profiles are useful tools for specifying the behaviour of onload.
However, they are only available when an application is accelerated using the
onload command line tool, not when using LD_PRELOAD.

Since each profile is used to set/unset environment variables, the behaviour
can be replicated by updating the environment of a user pod using a ConfigMap.

kustomize is used to patch the application pod without requiring any manual
changes.