ON-14911: Get container ID through controller runtime API
For crictl to start the cplane it needs to get the onload cplane
container's ID. The current implementation gets the container ID by
checking the cgroup in the container. Apart from being hacky, this
solution only worked for cgroup v1.

This change adds a new golang program which communicates with the
kuberenetes controller-runtime to get the details of the currently
running pod and to update the cplane server parameters. This will work
on both cgroup v1 and v2 systems.