ON-15027: ON-14984: Add kustomizations to build and deploy sfptpd
This change is to help prepare for supporting deployment using pre-built
binary container images. This introduces some new kustomization yaml
files to the sfptpd directory, which should make the deploying stuff to
cluster easier as well as allowing for a separation of the build step.

The kustomization yaml file in sfptpd/ will create an imagestream, make
a BuildConfig and create the DaemonSet. If a user wishes to one of these
steps separately then they can use the appropriate kustomization files
in either sfptpd/build or sfptpd/deploy.