ON-15027: Separate build and deploy steps for example pods
This changes adds extra kustomization files for building cns-sfnettest
without starting the user pods, and for starting the pods without
building an image.

Behaviour when applying the kustomization files in existing locations
should be unchanged.

If the user wants to deploy the example pods without building, they have
to create the imagestream manually before pushing the images.