ON-15022: Add onload_mkimages script
onload_mkimages.sh takes as input a filepath to a tarball release of
onload as well as some information about the cluster to build for
(OpenShift version, node kernel version) and produces a set of images
from the onload release that should work in the described cluster.

A new set of Dockerfile is also included in the scripts/ directory which
are used by onload_mkimages because the existing Dockerfiles are built
to pull from github, not work with local files.