ON-14945: Add kustomization to configure source locations
Adds a cluster/ directory with an onload-sources.conf file which
specificies the locations of tarballs for the Dockerfiles to pull and
build from. This defaults to using github as the source, but this change
will allow for easy configuration if github is inaccesible.

Also adds kustomization replacements to update the buildArgs for the
onload and sfc builds. Existing command line invocations should be
unaffected by this change.