ON-14911: Add onload-worker container to Device Plugin DaemonSet
Add the new "onload-worker" container to the existing Device Plugin
DaemonSet.

The new container provides the Onload control plane binary to the
Onload kernel instead of the dedicated control plane DaemonSet, which
the next patch will remove. Additionally, it manages (chcon and rm)
the Onload files in the host. Hence its name is "onload-worker".

Unlike the control plane DaemonSet, the new container uses the Golang
API to get the container ID instead of the cgroup workaround.

The new KubernetesHelper::GetContainerID() resides in the "pkg"
directory and is suitable for reuse.

Co-Authored-By: Thomas Crawley <thomas.crawley@xilinx.com>