ON-15122: Add upgrade logic for Onload CRs
This patch adds logic to the operator so that it can upgrade (or
downgrade) the version of onload used in the cluster.

The upgrade process is designed to work with KMM's ordered upgrade
process for kernel modules, but with additions so that it can control
the lifetimes of the other operands. Upgrades take place in a "rolling"
manner, the operator will follow the upgrade procedure for one node
before starting on the next node.

Labels are used to determine the lifetime of the operand's pod on the
node, currently the operator uses two labels: one for the module
lifetime and one for the device plugin daemonset. This does expose some
of the "innards" of the operator, but it should also help recovery if
the controller crashes.

There are some other changes in this patch to the length of the
controller's reconciliation loop. Previously the controller would try to
create each operand sequentially and only exit if everything was done or
if there was an error. However, due to the amount of time that an
upgrade can take I didn't think an approach using "long reconciliations"
would be practical. So I have changed the controller so that it will
requeue a reconciliation each time it would change the state of the
cluster.