ON-15122: Bump golang version to 1.21
Previously both the controller and the device plugin were using old
versions of golang (1.19 and 1.20 respectively), this prevented us from
using new features such as the "slices" package added to the standard
library in 1.21. Although, 1.21 features are only currently used in the
controller I have updated the dockerfiles for both images to use
golang:1.21 as the builder image so that we have fewer dependencies to
keep track of.

Also ran `go mod tidy`