ON-15122: Use field selector to list pods on node
Previously we would list all pods then filter manually, which isn't
ideal. This change adds an index which keeps track of pods and their
spec.nodeName which can be used as a field selector. This simplifies the
logic of the controller and reduces the amount of stuff we have to deal
with.

Unfortunately, it is not possible to use a more complicated field
selector to get pods using onload on a specified node. This is an
existing kubernetes limitation.