ON-15217: Create Module kind for SFC kernel module
The Onload operator controller will create the SFC module kind
if the SFC field is present in the Onload CR. In this case, the
SFC module kind will reuse the Onload's kernel image.

The controller also takes care of the upgrade of the SFC module
kind along with Onload. Moreover, if the old Onload CR does not
contain SFC but the new Onload CR does, it will create the new
module kind; if vice-versa, it will delete the SFC module kind.

Reviewed-by: Peter Colledge <peter.colledge@amd.com>
Reviewed-by: Thomas Crawley <thomas.crawley@amd.com>