SCL-1976: Update ContainerFileNotice
The operator controller and device plugin daemon set operate together
and implement the core product function, hence consolidating them in
one ContainerFileNotice.

The bundle image is an orthogonal and optional function, which we could
move into a sub-directory and create another ContainerFileNotice for it,
but since it is a self-contained image that does not download anything
from the Internet, this patch does not create ContainerFileNotice for
the bundle image.