ON-15420: Add module build options to onload CRD
Adds a subset of the KMM build options to the OnloadKernelMapping
struct. If present these options are converted in a KMM compatible form
and added to the Module struct. This allows the KMM Operator to perform
an in-cluster build.

The build field is cleared for the sfc module as otherwise two identical
builds would start on the cluster which would both be trying to write
the same image. It safe to clear the sfc definition and leave the onload
module untouched as it is not currently possible for the operator to
create a sfc module kind without a corresponding onload module kind, and
since they use the same images it will be able to start the pod when the
image is present.