ON-15420: Add in-cluster build example
This commit adds an example of how to use an in-cluster build, this is
done using kustomize to patch the onload CR definition and add a build
section which points at the onload-module-dockerfile added previously.

This change necessitated moving the sample onload CR yaml file into the
new config/samples/base directory to prevent kustomize complaining about
a circular dependency.

To create the onload-sample without the in-cluster build you should run:
    kubectl create -k config/samples
OR
    kubectl create -k config/samples/base

and to create it with the in-cluster build you should run:
    kubectl create -k config/samples/in-cluster-build