ON-14903: Add needNic flag to device plugin to allow for unit testing
Previously the device plugin would exit if no sfc nics were present on
the machine, this implementation limited the ability to unit test the
device plugin as unit tests should be able to run independently of the
hardware present on the machine.

If the flag is set to false it is possible to create a NicManager struct
on a machine without a nic present. This change was primarily made to
allow for unit tests, but theoretically could be used for other purposes
in the future.