ON-15386: Add options to configure onload file locations
Previously the location of onload files in the host and destination were
fixed constants, this isn't ideal if we want to support different
environments. This change adds new command line options to specify the
base onload location on the host as well as the destination locations
for libraries and binaries.

Also added a new unit test to test which libraries are mounted.