ON-15311: Add script to generate configmap from profile
Adds a new script `profile_to_configmap.sh` which is essentially a stripped
down version of the `onload` script that only includes the logic used to
process the profiles. The new script has the same syntax as the original
onload version, but now instead of setting environment variables it will
output the yaml text definition of a ConfigMap which can be used to set
environment variables in a user's pod.

Currently the script produces ConfigMaps with a fixed naming structure,
for example if you want to create a ConfigMap from a profile called
`name.opf` the generated name will be `onload-name-profile`.

If you want to apply the generated ConfigMap straight away, then you can
simply run:
`./profile_to_configmap.sh -p /path/to/profile.opf | oc apply -f -`

This patch also includes an example latency profile in config/samples/profiles