Set Device Plugin Image at controller level
Previously the device plugin image would be set per Onload CR, this
could potentially lead to confusion since (we expect) most users won't
need to change the device plugin image nor have fine-grained control of
the image. Additionally, since we have built the controller with certain
assumptions of the device plugin image it should lower the chances that
a user sets this to an incompatible value.