Cleanup: Refactor envtest cluster creation
Previously the code to start/stop an envtest cluster only existed in the
BeforeSuite/AfterSuite functions, which meant that all tests in the same
suite would use the same cluster. This change moves this code into some
new functions that can allow for the creation of additional clusters in
the same suite.