ON-15452: Add initial unit test for a dirty cluster
Unfortunately, we can't use the existing envtest infrastructure since it
starts the onload reconciler when the cluster is created thus there is
no time in which we can make it "dirty".

This change adds a new container for unit tests which will create a new
envtest cluster for each test without starting the onload reconciler.

As an initial test it will create the onload CR and the device plugin
before starting the reconciler and then check that it wasn't
deleted/re-created.