ON-15452: Add more unit tests for "dirty" environments
Adds two new tests that ensure the operands are updated as expected (but
not removed and re-added).

Also adds a new type of test that adds a Node to the envtest cluster.
These tests test how the controller handles starting when the operands
and node labels might not be as expected.