Cleanup: Run unit tests with ginkgo
As well as being a test framework, ginkgo can be used as an application
to run the test. It will run the same tests as `go test`, but ginkgo
allows tests within a suite to be run in concurrently (`go test` can
run suites in parallel, but each test in a suite sequentially). Since
the majority of our tests are in the same suite (controllers/) it can
save time running with ginkgo directly