ON-15180: Read sysfs to check for NICs
Rather than rely on an external tool (lshw) which requires extra
dependencies in container images and doesn't provide an simple way of
determining the vendor of a nic, it would be better to search through
sysfs directly (since that is basically all lshw is doing). This
approach also allows us to easily get other information from sysfs (such
as numa nodes).